requirements.txt
# Core dependencies
langchain>=0.1.0
streamlit>=1.31.0
pydantic>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# LLM providers
anthropic>=0.8.0
openai>=1.10.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-watch>=4.2.0
pytest-asyncio>=0.21.0
pynguin>=0.29.0

# Utilities
rich>=13.5.0
loguru>=0.7.0
# Note: streamlit-markdown-editor was removed as it's not available in PyPI
